+++ client/src/App.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react'
-import Registration from './Pages/registration/'
+import Registration from './Pages/Registration'
 
 class App extends Component {
   render() {
     return (
         <Registration />
-    );
+    )
   }
 }
 
-export default App;
+export default App
