+++ client/src/Pages/registration/styles.css
@@ -313,6 +313,7 @@ p {
         height: 100vh;
         width: 100vw;
     }
+
     .regContainer {
         display: flex;
         flex-direction: column;
@@ -320,17 +321,21 @@ p {
         width: 100%;
         height: 100%;
     }
+
     .formContainer {
         height: 100%;
         width: 100%;
     }
+
     .Logo {
         display: none;
         background: none;
     }
+
     #mobileLogo {
         display: flex;
     }
+    
     .logoStyle {
         width: 80%;
         margin-left: auto;
